+++ pub/telegram/cmd.sh
@@ -1,7 +1,8 @@
 ##? Developer Commands
 # flutter create --template = package cmd --org=usama.dev
-cd flutter/example
-flutter pub run cmd:install 
+# cd flutter/example
+# flutter pub run cmd:install 
+
 flutter pub publish --dry-run
 flutter pub publish
 flutter pub publish --force