+++ pub/telegram/CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.0.8
+📨 Send Message via Telegram
+
+🔗 Telegram Link Generator
+
+📌 Copy Telegram Link
+
+💬 Open Chat Directly
+
+📢 Join Telegram Channel/Group
+
+✅ Check if Telegram is Installed
+
+📞 Share Contact via Telegram
+
+👥 Open Telegram Group
+
+📷 Send Media via Telegram
+
+🔍 Check Username Availability
+
+🤖 Open Telegram Bot
+
+🔥 _more coming soon..._
+
 ## 0.0.7
 Minor Updates
 ## 0.0.6